Create Sample.java
Sample class test for SQLite Database.
While transferring to another environment, need to add an empty `tst.db` file.
Also, install SQLite JDBC jar file with classpath. For a building with an IDE, you need to configure the classpath variable from the project IDE with the name of the class. For example, the variable name is `org.sqlite.JDBC`, and the path is the jar class archive.
Otherwise, we can build the command prompt and configure the classpath from the prompt.